make font size respect browser settings
The font size in my browser is set to 20px (I am blind, I know), but due
to the font size being hardcoded as a constant pixel value in the css, my
settings are overwritten and the text appears small.

This commit mitigates that issue by replacing all fixed size font sizes
with rem units that depend on the browser-provided font size.